refactor: Consolidate authentication logic into useSession hook and improve error handling in login flow
