+++ src/hooks/useSession.ts
@@ -1,14 +1,13 @@
 import { useEffect } from "react";
 import { useCookies } from "react-cookie";
-import { useJwt } from "react-jwt";
+import { decodeToken, useJwt } from "react-jwt";
+import { toast } from "sonner";
 import { Session, User } from "@/types/session.types";
-import { COOKIE_AUTH_NAME } from "@/lib/auth/constants";
+import { COOKIE_AUTH_NAME, COOKIE_OPTIONS } from "@/lib/auth/constants";
 
 export default function useSession(): Session {
-    const [cookies] = useCookies([COOKIE_AUTH_NAME]);
-    const { decodedToken, isExpired, reEvaluateToken } = useJwt<User>(
-        cookies[COOKIE_AUTH_NAME],
-    );
+    const [cookies, setCookie, removeCookie] = useCookies([COOKIE_AUTH_NAME]);
+    const { decodedToken, isExpired, reEvaluateToken } = useJwt<User>(cookies[COOKIE_AUTH_NAME]);
 
     useEffect(() => {
         const sessionToken = cookies[COOKIE_AUTH_NAME];
@@ -16,14 +15,35 @@ export default function useSession(): Session {
             return;
         }
 
-        if (sessionToken && !decodedToken) {
+        if (sessionToken && (!decodedToken || isExpired)) {
+            removeCookie(COOKIE_AUTH_NAME, COOKIE_OPTIONS);
             reEvaluateToken(sessionToken);
+            toast.error("Session expired. Please login again.");
+            return;
         }
     }, [cookies, decodedToken, reEvaluateToken]);
 
+    const login = (token: string) => {
+        const decodedToken = decodeToken<User>(token);
+
+        if (!decodedToken) {
+            toast.error("Invalid token");
+            return null;
+        }
+
+        setCookie(COOKIE_AUTH_NAME, token, { ...COOKIE_OPTIONS, expires: new Date(decodedToken.exp * 1000) });
+        reEvaluateToken(token);
+        return decodedToken;
+    };
+
+    const logout = () => {
+        removeCookie(COOKIE_AUTH_NAME, COOKIE_OPTIONS);
+        reEvaluateToken("");
+    };
+
     if (!decodedToken || isExpired) {
-        return { data: null };
+        return { data: null, login, logout };
     }
 
-    return { data: decodedToken };
+    return { data: decodedToken, login, logout };
 }
