+++ src/lib/auth/constants.ts
@@ -3,5 +3,4 @@ export const COOKIE_OPTIONS = {
     path: "/",
     secure: process.env.NODE_ENV === "production",
     sameSite: "lax" as const,
-    maxAge: 0,
 };
